Add recipe for carbonara
In this commit I have added a recipe for a carbonara. This has included updating the index.html so that it links to the carbonara page, creating a directory within the main directory for recipes, adding a carbonara recipe to this, and also creating a directory within the main for images, where I will store all images.