Update index.html to include unordered list
I have updated the index file so that now the links to each recipes will be on a separate link with a bullet point before each one. Note that only the carbonara recipe has been included for now and I will add other recipes at a later date with a new commit.